Refactor packet struct related stuff to packet.go file
