+++ io.go
@@ -23,28 +23,6 @@ type FileHeader struct {
 	LinkType     uint32
 }
 
-type PacketTime struct {
-	Sec  int32
-	Usec int32
-}
-
-// Packet is a single packet parsed from a pcap file.
-type Packet struct {
-	// porting from 'pcap_pkthdr' struct
-	Time   time.Time // packet send/receive time
-	Caplen uint32    // bytes stored in the file (caplen <= len)
-	Len    uint32    // bytes sent/received
-
-	Data   []byte    // packet data
-
-	Type    int // protocol type, see LINKTYPE_*
-	DestMac uint64
-	SrcMac  uint64
-
-	Headers []interface{} // decoded headers, in order
-	Payload []byte        // remaining non-header bytes
-}
-
 // Reader parses pcap files.
 type Reader struct {
 	flip         bool
